Add configuration management and restructure bot modules
- Introduced a new `configs.py` file for centralized configuration management, replacing the previous `app/configs.py`.
- Updated `main.py` to utilize the new configuration structure for bot token and database settings.
- Removed the old `app/__init__.py` and `app/configs.py` files to streamline the codebase.
- Created new modules in the `bot` directory, including handlers, filters, and utilities, to enhance modularity and organization.
- Implemented various bot handlers for menu, language selection, search functionality, and track management, improving user interaction and experience.
- Added middleware for authentication and internationalization, ensuring user data is managed effectively and language preferences are respected.

This update enhances the bot's architecture, making it more maintainable and scalable for future features.
