--- app/__init__.py
@@ -1,5 +1,0 @@
-"""App init module."""
-from .configs import app_config
-
-
-__all__ = ['app_config']
