+++ bot/filters/__init__.py
@@ -1,4 +1,4 @@
-"""Filters for the app."""
+"""Filters for the bot."""
 from .language import LanguageFilter
 
 __all__ = ['LanguageFilter']
