+++ database/engine.py
@@ -3,12 +3,12 @@ import logging
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
 from sqlalchemy.orm import sessionmaker
-from app import app_config
+from configs import db_config
 
 
 logger = logging.getLogger(__name__)
 
-engine = create_async_engine(app_config.db_url, future=True)
+engine = create_async_engine(db_config.url, future=True)
 Base = declarative_base()
 
 async_session_factory = sessionmaker(
